requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=py38h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cloudpickle=3.0.0=pypi_0
contourpy=1.1.1=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.6.68=0
cuda-runtime=12.1.0=0
cuda-version=12.6=3
cycler=0.12.1=pypi_0
et-xmlfile=1.1.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
fonttools=4.53.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.9.0=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.34.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grad-cam=1.5.3=pypi_0
grpcio=1.66.1=pypi_0
idna=3.7=py38h06a4308_0
importlib-metadata=8.4.0=pypi_0
importlib-resources=6.4.4=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py38h06a4308_0
joblib=1.4.2=py38h06a4308_0
jpeg=9e=h5eee18b_3
kiwisolver=1.4.7=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.11.1.6=0
libcurand=10.3.7.68=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
lightning-utilities=0.9.0=py38h06a4308_0
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.41.1=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.7=pypi_0
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.7.5=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
numba=0.58.1=pypi_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
oauthlib=3.2.2=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openpyxl=3.1.5=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.1=py38h06a4308_0
pandas=2.0.3=pypi_0
pillow=10.4.0=py38h5eee18b_0
pip=24.2=py38h06a4308_0
platformdirs=3.10.0=py38h06a4308_0
pooch=1.7.0=py38h06a4308_0
protobuf=5.28.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pyparsing=3.1.4=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.4.1=py3.8_cuda12.1_cudnn9.1.0_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py38h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.32.3=py38h06a4308_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.3.0=py38h1128e8f_1
scipy=1.10.1=py38hf6e8229_1
seaborn=0.13.2=pypi_0
setuptools=72.1.0=py38h06a4308_0
shap=0.44.1=pypi_0
six=1.16.0=pypi_0
slicer=0.0.7=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.2=py38h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.12.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.5.0=py38h2f386ee_0
tk=8.6.14=h39e8969_0
torchaudio=2.4.1=py38_cu121
torchmetrics=1.4.0.post0=py38h06a4308_0
torchtriton=3.0.0=py38
torchvision=0.19.1=py38_cu121
tqdm=4.66.5=pypi_0
ttach=0.0.3=pypi_0
typing-extensions=4.11.0=py38h06a4308_0
typing_extensions=4.11.0=py38h06a4308_0
tzdata=2024.1=pypi_0
urllib3=2.2.2=py38h06a4308_0
werkzeug=3.0.4=pypi_0
wheel=0.43.0=py38h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zipp=3.20.1=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
